Separate StoreError into separate module
